requirements.txt
Flask==2.2.5
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Werkzeug==2.2.3


gunicorn==20.1.0
flask-cors==3.0.10


numpy==1.23.5
scipy==1.10.1
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
transformers==4.28.1
tensorflow==2.12.0



SQLAlchemy==2.0.15
psycopg2-binary==2.9.6
# sqlite==3.39.4


bcrypt==4.0.1
passlib==1.7.4
itsdangerous==2.1.2


pytest==7.3.1
pytest-flask==1.2.0
black==23.3.0
flake8==6.0.0
mypy==1.2.0
pre-commit==3.3.1


Click==8.1.3
python-dotenv==1.0.0
pydantic==1.10.7
requests==2.28.2
aiohttp==3.8.4